refactor of validation to do less passes
