--- src/parsers/strategies_parser.rs
@@ -1,229 +1,0 @@
-use crate::parsers::strategies::Strategies;
-use crate::parsers::strategy_structs::*;
-
-pub fn parse(
-    strategies: Vec<StrategyInFile>,
-    transformer_overrides: TransformerOverrides,
-) -> Strategies {
-    let mut transformed_strategies = Strategies::new();
-    //TODO If all columns are none, lets not do any transforming?
-    for strategy in strategies {
-        let columns = strategy
-            .columns
-            .into_iter()
-            .map(|column| {
-                (
-                    column.name.clone(),
-                    ColumnInfo {
-                        data_category: column.data_category.clone(),
-                        name: column.name.clone(),
-                        transformer: transformer(column, &transformer_overrides),
-                    },
-                )
-            })
-            .collect();
-
-        transformed_strategies.insert(strategy.table_name, columns);
-    }
-
-    transformed_strategies
-}
-
-fn column_transformer_is_overriden(
-    data_category: DataCategory,
-    overrides: &TransformerOverrides,
-) -> bool {
-    (data_category == DataCategory::PotentialPii && overrides.allow_potential_pii)
-        || (data_category == DataCategory::CommerciallySensitive
-            && overrides.allow_commercially_sensitive)
-}
-fn transformer(column: ColumnInFile, overrides: &TransformerOverrides) -> Transformer {
-    if column_transformer_is_overriden(column.data_category, overrides) {
-        Transformer {
-            name: TransformerType::Identity,
-            args: None,
-        }
-    } else {
-        column.transformer
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-
-    const TABLE_NAME: &str = "gert_lush_table";
-    const PII_COLUMN_NAME: &str = "pii_column";
-    const COMMERCIALLY_SENSITIVE_COLUMN_NAME: &str = "commercially_sensitive_column";
-
-    #[test]
-    fn can_parse_file_contents_into_hashmaps() {
-        let column_name = "column1";
-
-        let strategies = vec![StrategyInFile {
-            table_name: TABLE_NAME.to_string(),
-            description: "description".to_string(),
-            columns: vec![column_in_file(
-                DataCategory::Pii,
-                column_name,
-                TransformerType::Scramble,
-            )],
-        }];
-
-        let expected = Strategies::new_from(
-            TABLE_NAME.to_string(),
-            HashMap::from([(
-                column_name.to_string(),
-                ColumnInfo::builder()
-                    .with_name(column_name)
-                    .with_data_category(DataCategory::Pii)
-                    .with_transformer(TransformerType::Scramble, None)
-                    .build(),
-            )]),
-        );
-        let parsed = parse(strategies, TransformerOverrides::none());
-        assert_eq!(expected, parsed);
-    }
-
-    #[test]
-    fn ignores_transformers_for_potential_pii_if_flag_provided() {
-        let strategies = vec![StrategyInFile {
-            table_name: TABLE_NAME.to_string(),
-            description: "description".to_string(),
-            columns: vec![
-                column_in_file(
-                    DataCategory::PotentialPii,
-                    PII_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-                column_in_file(
-                    DataCategory::CommerciallySensitive,
-                    COMMERCIALLY_SENSITIVE_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-            ],
-        }];
-
-        let parsed = parse(
-            strategies,
-            TransformerOverrides {
-                allow_potential_pii: true,
-                allow_commercially_sensitive: false,
-            },
-        );
-        let pii_column_transformer = transformer_for_column(PII_COLUMN_NAME, &parsed);
-        let commercially_sensitive_transformer =
-            transformer_for_column(COMMERCIALLY_SENSITIVE_COLUMN_NAME, &parsed);
-
-        assert_eq!(pii_column_transformer.name, TransformerType::Identity);
-        assert_eq!(pii_column_transformer.args, None);
-
-        assert_eq!(
-            commercially_sensitive_transformer.name,
-            TransformerType::Scramble
-        );
-        assert_eq!(commercially_sensitive_transformer.args, None);
-    }
-
-    #[test]
-    fn ignores_transformers_for_commercially_sensitive_if_flag_provided() {
-        let strategies = vec![StrategyInFile {
-            table_name: TABLE_NAME.to_string(),
-            description: "description".to_string(),
-            columns: vec![
-                column_in_file(
-                    DataCategory::PotentialPii,
-                    PII_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-                column_in_file(
-                    DataCategory::CommerciallySensitive,
-                    COMMERCIALLY_SENSITIVE_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-            ],
-        }];
-
-        let parsed = parse(
-            strategies,
-            TransformerOverrides {
-                allow_potential_pii: false,
-                allow_commercially_sensitive: true,
-            },
-        );
-
-        let commercially_sensitive_transformer =
-            transformer_for_column(COMMERCIALLY_SENSITIVE_COLUMN_NAME, &parsed);
-        let pii_column_transformer = transformer_for_column(PII_COLUMN_NAME, &parsed);
-
-        assert_eq!(
-            commercially_sensitive_transformer.name,
-            TransformerType::Identity
-        );
-        assert_eq!(commercially_sensitive_transformer.args, None);
-
-        assert_eq!(pii_column_transformer.name, TransformerType::Scramble);
-        assert_eq!(pii_column_transformer.args, None);
-    }
-
-    #[test]
-    fn can_combine_override_flags() {
-        let strategies = vec![StrategyInFile {
-            table_name: TABLE_NAME.to_string(),
-            description: "description".to_string(),
-            columns: vec![
-                column_in_file(
-                    DataCategory::PotentialPii,
-                    PII_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-                column_in_file(
-                    DataCategory::CommerciallySensitive,
-                    COMMERCIALLY_SENSITIVE_COLUMN_NAME,
-                    TransformerType::Scramble,
-                ),
-            ],
-        }];
-
-        let parsed = parse(
-            strategies,
-            TransformerOverrides {
-                allow_potential_pii: true,
-                allow_commercially_sensitive: true,
-            },
-        );
-
-        let commercially_sensitive_transformer =
-            transformer_for_column(COMMERCIALLY_SENSITIVE_COLUMN_NAME, &parsed);
-        let pii_column_transformer = transformer_for_column(PII_COLUMN_NAME, &parsed);
-
-        assert_eq!(
-            commercially_sensitive_transformer.name,
-            TransformerType::Identity
-        );
-        assert_eq!(pii_column_transformer.name, TransformerType::Identity);
-    }
-
-    fn transformer_for_column(column_name: &str, strategies: &Strategies) -> Transformer {
-        strategies
-            .transformer_for_column(TABLE_NAME, column_name)
-            .expect("expecting a transformer!")
-    }
-
-    fn column_in_file(
-        data_category: DataCategory,
-        name: &str,
-        transformer_type: TransformerType,
-    ) -> ColumnInFile {
-        ColumnInFile {
-            data_category,
-            description: name.to_string(),
-            name: name.to_string(),
-            transformer: Transformer {
-                name: transformer_type,
-                args: None,
-            },
-        }
-    }
-}
